# contracts/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Egypt Industrial Initiatives Platform API",
  "version": "1.1.0"
 },
 "servers": [
  {
   "url": "http://localhost:4000"
  },
  {
   "url": "https://api.industry.gov.eg"
  }
 ],
 "paths": {
...
  "/api/v1/applications/{id}/messages": {
   "get": {
    "summary": "List application messages",
    "responses": {
     "200": {
      "description": "ok"
     },
     "403": {
      "description": "forbidden"
     },
     "404": {
      "description": "not found"
     }
    }
   },
   "post": {
    "summary": "Send application message",
    "responses": {
     "201": {
      "description": "created"
     },
     "400": {
      "description": "validation error"
     },
     "403": {
      "description": "forbidden"
     },
     "404": {
      "description": "not found"
     },
     "429": {
      "description": "rate limited"
     }
    }
   }
  },
...
  "/api/v1/audit-logs/export": {
   "post": {
    "summary": "Log a client-side export in the audit log { summaryAr }",
    "responses": {
     "201": {
      "description": "logged"
     },
     "400": {
      "description": "validation error"
     },
     "403": {
      "description": "forbidden"
     }
    }
   }
  },
...
 }
}